% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alchemer_surveys.R
\name{alchemer_surveys}
\alias{alchemer_surveys}
\title{Viewing Alchemer Surveys}
\usage{
alchemer_surveys(survey_ids)
}
\arguments{
\item{survey_ids}{optional argument of class character vector. If bp_surveys is false you must enter a character in this argument on which survey ID numbers you are interested in loading data for.}
}
\value{
data frame of list of all surveys available on Alchemer.
}
\description{
Viewing Alchemer Surveys
}
\examples{
data = alchemer_surveys(survey_ids = c("6726093","7311939"))
data = alchemer_surveys(survey_ids = "7311939")
}
