% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.R
\name{co}
\alias{co}
\title{Git Commit \emph{NPC Style}}
\usage{
co()
}
\value{
summary of the five most recent commits
}
\description{
Automatically commits and updates any changes you have saved to the pdrive and git hub repository. You will be asked to enter your commit message into the consol and press Enter.
}
\examples{
> co()
> Enter commit message: sc
"small changes"

> Enter commit message: ls

> Enter commit message:
> ""
> "No commits made or files staged"

}
