% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_font_palette.R
\name{color_font_palette}
\alias{color_font_palette}
\title{Function for making custom reports with NPC fonts and colors}
\usage{
color_font_palette()
}
\value{
list of element c which contains hex color codes and element f which contains ftext fonts
}
\description{
Function for making custom reports with NPC fonts and colors
}
