% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_airtable_rows.R
\name{create_airtable_rows}
\alias{create_airtable_rows}
\title{Function for Creating AirTable Rows}
\usage{
create_airtable_rows(
  data,
  table,
  base,
  single_select_fields = character(),
  link_fields = character(),
  first_row_i,
  last_row_i
)
}
\arguments{
\item{data}{A class of "string" that is the AirTable Base you are uploading data to.}

\item{table}{A class of "data.frame" that is the name of table I am wanting to upload to AirTable.}

\item{base}{A class of "string" that is the AirTable Base you are uploading data to.}

\item{single_select_fields}{A class of "character vector" that is a list of the single select columns.}

\item{link_fields}{A class of "character vector" that is a list of the class "url" columns.}
}
\value{
HTTPs response on if the rows were uploaded successfully or not. Message 200 means that the rows were uploaded correctly.
}
\description{
Function for Creating AirTable Rows
}
\examples{

row_indices = data.frame(
 first_row_i = seq(1, nrow(df), 10),
 last_row_i  = c(seq(min(10, nrow(df)), nrow(df), 10),nrow(df))
 )

Sys.sleep(30)
sapply(
 1:nrow(row_indices),
 function(index) {
   create_airtable_rows(
     data = df,
     table = "Contact List",
     base = "Assessment Tracker",
     single_select_fields = character(),
     link_fields = character(),
     first_row_i = row_indices[index,"first_row_i"],
     last_row_i = row_indices[index,"last_row_i"])
 }
)
}
