% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bookmarks_html.R
\name{create_bookmarks_html}
\alias{create_bookmarks_html}
\title{create_bookmarks_html}
\usage{
create_bookmarks_html(urls, titles = NULL, file_path = NULL)
}
\arguments{
\item{urls}{a variable of class character vector of url links}

\item{titles}{an optional variable of class character vector of what you would like your bookmarks to be named}

\item{file_path}{an optional variable of class character that is a file path of where you want your html file to be saved to. Default is C:/Users/USER/Documents/bookmarks.html}
}
\value{
a saved file of an bookmarks as an html file you can upload to chrome
}
\description{
create_bookmarks_html
}
\examples{
urls = paste0("https://app.alchemer.com/explorer/export-summary/id/",10000:10001) # define surveys you want to open
create_bookmarks_html(urls, titles = c("Survey 1", "Survey 2"), file_path = "P:\...surveys.html")
}
