% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_missing_directory.R
\name{creating_missing_directory}
\alias{creating_missing_directory}
\title{Creating a missing Directory}
\usage{
creating_missing_directory(path)
}
\arguments{
\item{path}{argument of class character}
}
\value{
a print statement on the status of the directory you want created.
}
\description{
Creating a missing Directory
}
\examples{
creating_missing_directory(paste0("C:/Users/lastname/Documents/",Sys.Date()))
}
