% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_all_airtable_records.R
\name{delete_all_airtable_records}
\alias{delete_all_airtable_records}
\title{Function for Deleting all rows in an AirTable table}
\usage{
delete_all_airtable_records(table, base)
}
\arguments{
\item{base}{A class of "string" that is the AirTable Base you are deleting data from.}

\item{data}{A class of "string" that is the AirTable Base you are deleting data from.}
}
\value{
HTTPs response on if the rows were deleted successfully or not. Message 200 means that the rows were deleted correctly.
}
\description{
Function for Deleting all rows in an AirTable table
}
\examples{
delete_all_airtable_records(base = "Assessment Tracker", table = "Contact List")
}
