% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_airtable_ids.R
\name{finding_airtable_ids}
\alias{finding_airtable_ids}
\title{Function for Finding AirTable IDs}
\usage{
finding_airtable_ids(table, base = "Assessment Tracker")
}
\arguments{
\item{table}{A class of "string" that is the name of table I am interested in reading in.}

\item{base}{A class of "string" that is the AirTable Base with my table located in it.}
}
\value{
IDs for our Base and Table
}
\description{
Function for Finding AirTable IDs
}
\examples{
finding_airtable_ids(table = "Logins",base = "Assessment Tracker")
}
