% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitc.R
\name{gitc}
\alias{gitc}
\title{Git Commit \emph{NPC Style}}
\usage{
gitc(files = TRUE, beep = TRUE)
}
\arguments{
\item{files}{optional list argument of the of files you wish to commit changes to. Default is all files with changes saved.}
}
\value{
summary of the five most recent commits
}
\description{
Automatically commits and updates any changes you have saved to the pdrive and git hub repository. You will be asked to enter your commit message into the consol and press Enter.
}
\examples{
> gitc()
> Made report title larger font

> gitc("cleaning_data.R")
> finished joining tc data sets

}
