% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project}
\alias{initialize_project}
\title{initialize_project}
\usage{
initialize_project(mail_merge = TRUE)
}
\arguments{
\item{mail_merge}{a variable of class logical for specifying if we want a file names ./R/mail_merge.R}
}
\value{
directories and syntax written for specific files
}
\description{
initialize_project
}
