% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_style.R
\name{npc_style}
\alias{npc_style}
\title{NPC Research Custom Theme}
\usage{
npc_style()
}
\description{
This function allows you to easily and consistenly alter the style of your ggplot
data viz in the style of NPC Research.
}
\keyword{npc_style}
