% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnames.R
\name{pnames}
\alias{pnames}
\title{Function for printing strings in colsole in a easy to read format}
\usage{
pnames(input, var_names = TRUE)
}
\arguments{
\item{var_names}{A class of "logical", depending on what class our input value is}

\item{data}{A class of "data.frame" or "string" based on our second argument var_names}
}
\value{
output in console of string printed with line breaks
}
\description{
Function for printing strings in colsole in a easy to read format
}
\examples{
pnames(iris)
Sepal.Length
Sepal.Width
Petal.Length
Petal.Width
Species

pnames(fruit[1:3],F)
apple
apricot
avocado
}
