% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_word_cloud.R
\name{print_word_cloud}
\alias{print_word_cloud}
\title{Function for printing a basic word cloud}
\usage{
print_word_cloud(text)
}
\arguments{
\item{text}{A class of character or character vector}
}
\value{
a word cloud in the viewer pane
}
\description{
Function for printing a basic word cloud
}
\examples{
print_word_cloud(text = c("a","b",letters[1:13]))
}
