% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_site_info.R
\name{pull_site_info}
\alias{pull_site_info}
\title{Function for getting info on if a link is still valid or not}
\usage{
pull_site_info(url)
}
\arguments{
\item{url}{A class of "string" that is used to search through your default browser}
}
\value{
list with info about https request
}
\description{
Function for getting info on if a link is still valid or not
}
\examples{

pull_site_info("https://npcresearch.com/")
pull_site_info("https://httpbin.org/status/200")
pull_site_info("https://httpbin.org/redirect/1")
pull_site_info("https://httpbin.org/status/302")
pull_site_info("https://httpbin.org/status/401")
pull_site_info("https://httpbin.org/status/403")
pull_site_info("https://httpbin.org/status/404")
pull_site_info("https://httpbin.org/status/429")
pull_site_info("https://github.com")

tribble(
 ~name, ~link,
 "npc", "https://npcresearch.com/",
 "200", "https://httpbin.org/status/200",
 "401", "https://httpbin.org/status/401",
 "403", "https://httpbin.org/status/403",
 "404", "https://httpbin.org/status/404"
 ) \%>\%
 mutate(info = map(link, pull_site_info)) \%>\%
 tidyr::unnest_wider(info)
}
