% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_airtable.R
\name{read_airtable}
\alias{read_airtable}
\title{Function for reading in Air Table tables}
\usage{
read_airtable(table, base = "Assessment Tracker")
}
\arguments{
\item{table}{A class of "string" that is the name of table I am interested in reading in.}

\item{base}{A class of "string" that is the AirTable Base with my table located in it.}
}
\value{
Air Table table
}
\description{
Function for reading in Air Table tables
}
\examples{


assessments = read_airtable(table = "Assessments",base = "Assessment Tracker")
courts = read_airtable(table = "Courts",base = "Assessment Tracker")
survey_responses = read_airtable(table = "Survey Responses",base = "Assessment Tracker")
}
