% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_alchemer.R
\name{read_alchemer}
\alias{read_alchemer}
\title{Reading Alchemer Data}
\usage{
read_alchemer(
  survey_ids = 7589884,
  end_date = Sys.Date(),
  start_date = Sys.Date() - days(14)
)
}
\arguments{
\item{survey_ids}{argument of class character vector. If bp_surveys is false you must enter a character in this argument on which survey ID numbers you are interested in loading data for.}

\item{end_date}{argument of class string, that should be specified as a date for when you want to extract survey data}

\item{start_date}{argument of class string, that should be specified as a date for when you want to extract survey data}
}
\value{
data frame of selected survey's survey responses with survey name start date, submit date, reference number and assessment id.
}
\description{
Reading Alchemer Data
}
\examples{
data = read_alchemer(survey_ids = c("6726093","7311939"))
}
