% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss_text_vertical.R
\name{spss_text_vertical}
\alias{spss_text_vertical}
\title{Function for taking spss syntax and making it vertial with line breaks.}
\usage{
spss_text_vertical(file = "~/text.txt")
}
\arguments{
\item{file}{path to a text file.}
}
\value{
output of spss variable names with line broke after every name.
}
\description{
Function for taking spss syntax and making it vertial with line breaks.
}
\examples{

writeLines(text = "var1 var2
var3", con = "~/text.txt")

spss_text_vertical("~/text.txt")

var1
var2
var3
}
