% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_heat_map.R
\name{state_heat_map}
\alias{state_heat_map}
\title{Function for Creating Heat Map of US States}
\usage{
state_heat_map(state_data, legend_title = NULL)
}
\arguments{
\item{legend_title}{A class of "string" that is what you want the legend title to display.}

\item{data}{A class of data.frame that has variables state_abbr and count}
}
\value{
ggplot heat map
}
\description{
Function for Creating Heat Map of US States
}
\examples{

library(usmap)

data = haven::read_sav("P:/6. Projects Active/Harm Reduction in Tx Cts/Survey/Data/Alchemer Exports/Harm Reduction Survey_working.sav") \%>\%
sjlabelled::as_label() \%>\%
  select(state_abbr = Q1.3_4, type = Q1.1) \%>\%
  # filter(type == "Adult Treatment Court") \%>\%
  group_by(state_abbr) \%>\%
  summarise(count = n())

state_heat_map(data)

}
