% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_directory.R
\name{tree}
\alias{tree}
\title{Interactive Tree Map}
\usage{
tree(file_path, exclude, collapse)
}
\arguments{
\item{file_path}{string of a file path you have access to. It is
recommended to use forward slashes in between files. Two back slashes
is also an option.}

\item{exclude}{list of files you wish to exclude from the tree.}

\item{expand}{if FALSE, entire tree is expanded, Default is TRUE.}
}
\value{
Returns interactive html file
}
\description{
Outputs an interactive html file of a project directory
}
\examples{
path <- "P:/6. Projects Active/Example/Syntax"
tree(path)
tree("P:/6. Projects Active/Example", exclude = ".RData")
tree(
  file_path = "P:/6. Projects Active/Example/Data",
  exclude = c("data.csv", "data.sav"),
  expand = TRUE)

}
\references{
Adeel Khan
\pkg{collapsibleTree}:
\url{https://github.com/AdeelK93/collapsibleTree}
}
